refactor: split trace module by responsibility
