--- adl/src/trace.rs
@@ -1,1696 +1,0 @@
-//! Deterministic trace-event stream used across execution and replay tooling.
-//!
-//! `Trace` is the main in-memory event log for runtime execution, while
-//! `TraceEvent` captures concrete lifecycle, delegation, and step transitions
-//! used for artifact replay and audit.
-
-use std::collections::HashMap;
-use std::time::{Instant, SystemTime, UNIX_EPOCH};
-
-use crate::adl::DelegationSpec;
-use crate::execute::{ExecutionBoundary, RuntimeLifecyclePhase};
-
-/// In-memory execution trace captured during runtime operation.
-///
-/// A `Trace` stores high-level lifecycle markers and step-level telemetry in an
-/// append-only sequence, including timings for replay-friendly review.
-#[derive(Debug, Clone)]
-pub struct Trace {
-    pub run_id: String,
-    pub workflow_id: String,
-    pub version: String,
-    pub events: Vec<TraceEvent>,
-    run_started_ms: u128,
-    run_started_instant: Instant,
-    step_started_ms: HashMap<String, u128>,
-    delegation_ids: HashMap<String, String>,
-    next_delegation_counter: u64,
-}
-
-#[derive(Debug, Clone)]
-/// Canonical trace event model emitted by runtime execution.
-///
-/// Variants are intentionally stable and replay-oriented. Additional variants
-/// should preserve naming conventions and include bounded human-readable fields.
-pub enum TraceEvent {
-    LifecyclePhaseEntered {
-        ts_ms: u128,
-        elapsed_ms: u128,
-        phase: RuntimeLifecyclePhase,
-    },
-    ExecutionBoundaryCrossed {
-        ts_ms: u128,
-        elapsed_ms: u128,
-        boundary: ExecutionBoundary,
-        state: String,
-    },
-    GovernedProposalObserved {
-        ts_ms: u128,
-        elapsed_ms: u128,
-        proposal_id: String,
-        tool_name: String,
-        redacted_arguments_ref: String,
-    },
-    GovernedProposalNormalized {
-        ts_ms: u128,
-        elapsed_ms: u128,
-        proposal_id: String,
-        normalized_proposal_ref: String,
-        redacted_arguments_ref: String,
-    },
-    GovernedAccConstructed {
-        ts_ms: u128,
-        elapsed_ms: u128,
-        proposal_id: String,
-        acc_contract_id: String,
-        replay_posture: String,
-    },
-    GovernedPolicyInjected {
-        ts_ms: u128,
-        elapsed_ms: u128,
-        proposal_id: String,
-        policy_evidence_ref: String,
-        outcome: String,
-    },
-    GovernedVisibilityResolved {
-        ts_ms: u128,
-        elapsed_ms: u128,
-        proposal_id: String,
-        actor_view: String,
-        operator_view: String,
-        reviewer_view: String,
-        public_report_view: String,
-        observatory_projection: String,
-    },
-    GovernedFreedomGateDecided {
-        ts_ms: u128,
-        elapsed_ms: u128,
-        proposal_id: String,
-        candidate_id: String,
-        decision: String,
-        reason_code: String,
-        boundary: String,
-        redaction_summary: String,
-    },
-    GovernedActionSelected {
-        ts_ms: u128,
-        elapsed_ms: u128,
-        proposal_id: String,
-        action_id: String,
-        tool_name: String,
-        adapter_id: String,
-        evidence_refs: Vec<String>,
-    },
-    GovernedActionRejected {
-        ts_ms: u128,
-        elapsed_ms: u128,
-        proposal_id: String,
-        action_id: String,
-        tool_name: String,
-        adapter_id: String,
-        reason_code: String,
-        evidence_refs: Vec<String>,
-    },
-    GovernedExecutionResultRecorded {
-        ts_ms: u128,
-        elapsed_ms: u128,
-        proposal_id: String,
-        action_id: String,
-        adapter_id: String,
-        result_ref: String,
-        evidence_refs: Vec<String>,
-    },
-    GovernedRefusalRecorded {
-        ts_ms: u128,
-        elapsed_ms: u128,
-        proposal_id: String,
-        action_id: String,
-        reason_code: String,
-        evidence_refs: Vec<String>,
-    },
-    GovernedRedactionDecisionRecorded {
-        ts_ms: u128,
-        elapsed_ms: u128,
-        proposal_id: String,
-        audience: String,
-        surfaces: Vec<String>,
-        outcome: String,
-        detail: Option<String>,
-    },
-    SchedulerPolicy {
-        ts_ms: u128,
-        elapsed_ms: u128,
-        max_concurrency: usize,
-        source: String,
-    },
-    RunFailed {
-        ts_ms: u128,
-        elapsed_ms: u128,
-        message: String,
-    },
-    RunFinished {
-        ts_ms: u128,
-        elapsed_ms: u128,
-        success: bool,
-    },
-    StepStarted {
-        ts_ms: u128,
-        elapsed_ms: u128,
-        step_id: String,
-        agent_id: String,
-        provider_id: String,
-        task_id: String,
-        delegation: Option<DelegationSpec>,
-    },
-    PromptAssembled {
-        ts_ms: u128,
-        elapsed_ms: u128,
-        step_id: String,
-        prompt_hash: String,
-    },
-    StepOutputChunk {
-        ts_ms: u128,
-        elapsed_ms: u128,
-        step_id: String,
-        chunk_bytes: usize,
-    },
-    DelegationRequested {
-        ts_ms: u128,
-        elapsed_ms: u128,
-        delegation_id: String,
-        step_id: String,
-        action_kind: String,
-        target_id: String,
-    },
-    DelegationPolicyEvaluated {
-        ts_ms: u128,
-        elapsed_ms: u128,
-        delegation_id: String,
-        step_id: String,
-        action_kind: String,
-        target_id: String,
-        decision: String,
-        rule_id: Option<String>,
-    },
-    DelegationApproved {
-        ts_ms: u128,
-        elapsed_ms: u128,
-        delegation_id: String,
-        step_id: String,
-    },
-    DelegationDenied {
-        ts_ms: u128,
-        elapsed_ms: u128,
-        delegation_id: String,
-        step_id: String,
-        action_kind: String,
-        target_id: String,
-        rule_id: Option<String>,
-    },
-    DelegationDispatched {
-        ts_ms: u128,
-        elapsed_ms: u128,
-        delegation_id: String,
-        step_id: String,
-        action_kind: String,
-        target_id: String,
-    },
-    DelegationResultReceived {
-        ts_ms: u128,
-        elapsed_ms: u128,
-        delegation_id: String,
-        step_id: String,
-        success: bool,
-        output_bytes: usize,
-    },
-    DelegationCompleted {
-        ts_ms: u128,
-        elapsed_ms: u128,
-        delegation_id: String,
-        step_id: String,
-        outcome: String,
-    },
-    StepFinished {
-        ts_ms: u128,
-        elapsed_ms: u128,
-        step_id: String,
-        success: bool,
-        duration_ms: u128,
-    },
-    CallEntered {
-        ts_ms: u128,
-        elapsed_ms: u128,
-        caller_step_id: String,
-        callee_workflow_id: String,
-        namespace: String,
-    },
-    CallExited {
-        ts_ms: u128,
-        elapsed_ms: u128,
-        caller_step_id: String,
-        status: String,
-        namespace: String,
-    },
-}
-
-impl TraceEvent {
-    /// Convert an event to a short, stable summary string for logs and snapshots.
-    pub fn summarize(&self) -> String {
-        match self {
-            TraceEvent::LifecyclePhaseEntered {
-                ts_ms,
-                elapsed_ms,
-                phase,
-            } => format!(
-                "{} (+{}ms) LifecyclePhaseEntered phase={}",
-                format_ts_ms(*ts_ms),
-                elapsed_ms,
-                phase.as_str()
-            ),
-            TraceEvent::ExecutionBoundaryCrossed {
-                ts_ms,
-                elapsed_ms,
-                boundary,
-                state,
-            } => format!(
-                "{} (+{}ms) ExecutionBoundaryCrossed boundary={} state={}",
-                format_ts_ms(*ts_ms),
-                elapsed_ms,
-                boundary.as_str(),
-                state
-            ),
-            TraceEvent::GovernedProposalObserved {
-                ts_ms,
-                elapsed_ms,
-                proposal_id,
-                tool_name,
-                redacted_arguments_ref,
-            } => format!(
-                "{} (+{}ms) GovernedProposalObserved proposal_id={} tool={} arguments_ref={}",
-                format_ts_ms(*ts_ms),
-                elapsed_ms,
-                proposal_id,
-                tool_name,
-                redacted_arguments_ref
-            ),
-            TraceEvent::GovernedProposalNormalized {
-                ts_ms,
-                elapsed_ms,
-                proposal_id,
-                normalized_proposal_ref,
-                redacted_arguments_ref,
-            } => format!(
-                "{} (+{}ms) GovernedProposalNormalized proposal_id={} normalized_ref={} arguments_ref={}",
-                format_ts_ms(*ts_ms),
-                elapsed_ms,
-                proposal_id,
-                normalized_proposal_ref,
-                redacted_arguments_ref
-            ),
-            TraceEvent::GovernedAccConstructed {
-                ts_ms,
-                elapsed_ms,
-                proposal_id,
-                acc_contract_id,
-                replay_posture,
-            } => format!(
-                "{} (+{}ms) GovernedAccConstructed proposal_id={} acc={} replay_posture={}",
-                format_ts_ms(*ts_ms),
-                elapsed_ms,
-                proposal_id,
-                acc_contract_id,
-                replay_posture
-            ),
-            TraceEvent::GovernedPolicyInjected {
-                ts_ms,
-                elapsed_ms,
-                proposal_id,
-                policy_evidence_ref,
-                outcome,
-            } => format!(
-                "{} (+{}ms) GovernedPolicyInjected proposal_id={} policy_ref={} outcome={}",
-                format_ts_ms(*ts_ms),
-                elapsed_ms,
-                proposal_id,
-                policy_evidence_ref,
-                outcome
-            ),
-            TraceEvent::GovernedVisibilityResolved {
-                ts_ms,
-                elapsed_ms,
-                proposal_id,
-                actor_view,
-                operator_view,
-                reviewer_view,
-                public_report_view,
-                observatory_projection,
-            } => format!(
-                "{} (+{}ms) GovernedVisibilityResolved proposal_id={} actor_view={} operator_view={} reviewer_view={} public_view={} observatory={}",
-                format_ts_ms(*ts_ms),
-                elapsed_ms,
-                proposal_id,
-                actor_view,
-                operator_view,
-                reviewer_view,
-                public_report_view,
-                observatory_projection
-            ),
-            TraceEvent::GovernedFreedomGateDecided {
-                ts_ms,
-                elapsed_ms,
-                proposal_id,
-                candidate_id,
-                decision,
-                reason_code,
-                boundary,
-                redaction_summary,
-            } => format!(
-                "{} (+{}ms) GovernedFreedomGateDecided proposal_id={} candidate={} decision={} reason={} boundary={} redaction={}",
-                format_ts_ms(*ts_ms),
-                elapsed_ms,
-                proposal_id,
-                candidate_id,
-                decision,
-                reason_code,
-                boundary,
-                redaction_summary
-            ),
-            TraceEvent::GovernedActionSelected {
-                ts_ms,
-                elapsed_ms,
-                proposal_id,
-                action_id,
-                tool_name,
-                adapter_id,
-                evidence_refs,
-            } => format!(
-                "{} (+{}ms) GovernedActionSelected proposal_id={} action_id={} tool={} adapter={} evidence_count={}",
-                format_ts_ms(*ts_ms),
-                elapsed_ms,
-                proposal_id,
-                action_id,
-                tool_name,
-                adapter_id,
-                evidence_refs.len()
-            ),
-            TraceEvent::GovernedActionRejected {
-                ts_ms,
-                elapsed_ms,
-                proposal_id,
-                action_id,
-                tool_name,
-                adapter_id,
-                reason_code,
-                evidence_refs,
-            } => format!(
-                "{} (+{}ms) GovernedActionRejected proposal_id={} action_id={} tool={} adapter={} reason={} evidence_count={}",
-                format_ts_ms(*ts_ms),
-                elapsed_ms,
-                proposal_id,
-                action_id,
-                tool_name,
-                adapter_id,
-                reason_code,
-                evidence_refs.len()
-            ),
-            TraceEvent::GovernedExecutionResultRecorded {
-                ts_ms,
-                elapsed_ms,
-                proposal_id,
-                action_id,
-                adapter_id,
-                result_ref,
-                evidence_refs,
-            } => format!(
-                "{} (+{}ms) GovernedExecutionResultRecorded proposal_id={} action_id={} adapter={} result_ref={} evidence_count={}",
-                format_ts_ms(*ts_ms),
-                elapsed_ms,
-                proposal_id,
-                action_id,
-                adapter_id,
-                result_ref,
-                evidence_refs.len()
-            ),
-            TraceEvent::GovernedRefusalRecorded {
-                ts_ms,
-                elapsed_ms,
-                proposal_id,
-                action_id,
-                reason_code,
-                evidence_refs,
-            } => format!(
-                "{} (+{}ms) GovernedRefusalRecorded proposal_id={} action_id={} reason={} evidence_count={}",
-                format_ts_ms(*ts_ms),
-                elapsed_ms,
-                proposal_id,
-                action_id,
-                reason_code,
-                evidence_refs.len()
-            ),
-            TraceEvent::GovernedRedactionDecisionRecorded {
-                ts_ms,
-                elapsed_ms,
-                proposal_id,
-                audience,
-                surfaces,
-                outcome,
-                detail,
-            } => {
-                let surfaces = surfaces.join(",");
-                let base = format!(
-                    "{} (+{}ms) GovernedRedactionDecisionRecorded proposal_id={} audience={} surfaces={} outcome={}",
-                    format_ts_ms(*ts_ms),
-                    elapsed_ms,
-                    proposal_id,
-                    audience,
-                    surfaces,
-                    outcome
-                );
-                if let Some(detail) = detail {
-                    format!("{base} detail={detail}")
-                } else {
-                    base
-                }
-            }
-            TraceEvent::SchedulerPolicy {
-                ts_ms,
-                elapsed_ms,
-                max_concurrency,
-                source,
-            } => format!(
-                "{} (+{}ms) SchedulerPolicy max_concurrency={} source={}",
-                format_ts_ms(*ts_ms),
-                elapsed_ms,
-                max_concurrency,
-                source
-            ),
-            TraceEvent::RunFailed {
-                ts_ms,
-                elapsed_ms,
-                message,
-            } => format!(
-                "{} (+{}ms) RunFailed message={message}",
-                format_ts_ms(*ts_ms),
-                elapsed_ms
-            ),
-            TraceEvent::RunFinished {
-                ts_ms,
-                elapsed_ms,
-                success,
-            } => format!(
-                "{} (+{}ms) RunFinished success={success}",
-                format_ts_ms(*ts_ms),
-                elapsed_ms
-            ),
-            TraceEvent::StepStarted {
-                ts_ms,
-                elapsed_ms,
-                step_id,
-                agent_id,
-                provider_id,
-                task_id,
-                delegation,
-            } => {
-                let prefix = format!(
-                    "{} (+{}ms) StepStarted step={step_id} agent={agent_id} provider={provider_id} task={task_id}",
-                    format_ts_ms(*ts_ms),
-                    elapsed_ms
-                );
-                if let Some(json) = delegation_json(delegation.as_ref()) {
-                    format!("{prefix} delegation={json}")
-                } else {
-                    prefix
-                }
-            }
-            TraceEvent::PromptAssembled {
-                ts_ms,
-                elapsed_ms,
-                step_id,
-                prompt_hash,
-            } => format!(
-                "{} (+{}ms) PromptAssembled step={step_id} hash={prompt_hash}",
-                format_ts_ms(*ts_ms),
-                elapsed_ms
-            ),
-            TraceEvent::StepOutputChunk {
-                ts_ms,
-                elapsed_ms,
-                step_id,
-                chunk_bytes,
-            } => format!(
-                "{} (+{}ms) StepOutputChunk step={step_id} bytes={chunk_bytes}",
-                format_ts_ms(*ts_ms),
-                elapsed_ms
-            ),
-            TraceEvent::DelegationRequested {
-                ts_ms,
-                elapsed_ms,
-                delegation_id,
-                step_id,
-                action_kind,
-                target_id,
-            } => format!(
-                "{} (+{}ms) DelegationRequested delegation_id={} step={} action={} target={}",
-                format_ts_ms(*ts_ms),
-                elapsed_ms,
-                delegation_id,
-                step_id,
-                action_kind,
-                target_id
-            ),
-            TraceEvent::DelegationPolicyEvaluated {
-                ts_ms,
-                elapsed_ms,
-                delegation_id,
-                step_id,
-                action_kind,
-                target_id,
-                decision,
-                rule_id,
-            } => {
-                let base = format!(
-                    "{} (+{}ms) DelegationPolicyEvaluated delegation_id={} step={} action={} target={} decision={}",
-                    format_ts_ms(*ts_ms),
-                    elapsed_ms,
-                    delegation_id,
-                    step_id,
-                    action_kind,
-                    target_id,
-                    decision
-                );
-                if let Some(rule_id) = rule_id {
-                    format!("{base} rule_id={rule_id}")
-                } else {
-                    base
-                }
-            }
-            TraceEvent::DelegationApproved {
-                ts_ms,
-                elapsed_ms,
-                delegation_id,
-                step_id,
-            } => format!(
-                "{} (+{}ms) DelegationApproved delegation_id={} step={}",
-                format_ts_ms(*ts_ms),
-                elapsed_ms,
-                delegation_id,
-                step_id
-            ),
-            TraceEvent::DelegationDenied {
-                ts_ms,
-                elapsed_ms,
-                delegation_id,
-                step_id,
-                action_kind,
-                target_id,
-                rule_id,
-            } => {
-                let base = format!(
-                    "{} (+{}ms) DelegationDenied delegation_id={} step={} action={} target={}",
-                    format_ts_ms(*ts_ms),
-                    elapsed_ms,
-                    delegation_id,
-                    step_id,
-                    action_kind,
-                    target_id
-                );
-                if let Some(rule_id) = rule_id {
-                    format!("{base} rule_id={rule_id}")
-                } else {
-                    base
-                }
-            }
-            TraceEvent::DelegationDispatched {
-                ts_ms,
-                elapsed_ms,
-                delegation_id,
-                step_id,
-                action_kind,
-                target_id,
-            } => format!(
-                "{} (+{}ms) DelegationDispatched delegation_id={} step={} action={} target={}",
-                format_ts_ms(*ts_ms),
-                elapsed_ms,
-                delegation_id,
-                step_id,
-                action_kind,
-                target_id
-            ),
-            TraceEvent::DelegationResultReceived {
-                ts_ms,
-                elapsed_ms,
-                delegation_id,
-                step_id,
-                success,
-                output_bytes,
-            } => format!(
-                "{} (+{}ms) DelegationResultReceived delegation_id={} step={} success={} bytes={}",
-                format_ts_ms(*ts_ms),
-                elapsed_ms,
-                delegation_id,
-                step_id,
-                success,
-                output_bytes
-            ),
-            TraceEvent::DelegationCompleted {
-                ts_ms,
-                elapsed_ms,
-                delegation_id,
-                step_id,
-                outcome,
-            } => format!(
-                "{} (+{}ms) DelegationCompleted delegation_id={} step={} outcome={}",
-                format_ts_ms(*ts_ms),
-                elapsed_ms,
-                delegation_id,
-                step_id,
-                outcome
-            ),
-            TraceEvent::StepFinished {
-                ts_ms,
-                elapsed_ms,
-                step_id,
-                success,
-                duration_ms,
-            } => format!(
-                "{} (+{}ms) StepFinished step={step_id} success={success} duration={}",
-                format_ts_ms(*ts_ms),
-                elapsed_ms,
-                format_duration_secs(*duration_ms)
-            ),
-            TraceEvent::CallEntered {
-                ts_ms,
-                elapsed_ms,
-                caller_step_id,
-                callee_workflow_id,
-                namespace,
-            } => format!(
-                "{} (+{}ms) CallEntered caller_step={} callee_workflow={} namespace={}",
-                format_ts_ms(*ts_ms),
-                elapsed_ms,
-                caller_step_id,
-                callee_workflow_id,
-                namespace
-            ),
-            TraceEvent::CallExited {
-                ts_ms,
-                elapsed_ms,
-                caller_step_id,
-                status,
-                namespace,
-            } => format!(
-                "{} (+{}ms) CallExited caller_step={} status={} namespace={}",
-                format_ts_ms(*ts_ms),
-                elapsed_ms,
-                caller_step_id,
-                status,
-                namespace
-            ),
-        }
-    }
-}
-
-impl Trace {
-    /// Construct a new trace context for a run/workflow pair.
-    ///
-    /// `run_id` and `workflow_id` participate in replay artifact naming and
-    /// are emitted verbatim in trace output.
-    pub fn new(
-        run_id: impl Into<String>,
-        workflow_id: impl Into<String>,
-        version: impl Into<String>,
-    ) -> Self {
-        let started = Self::now_ms();
-        Self {
-            run_id: run_id.into(),
-            workflow_id: workflow_id.into(),
-            version: version.into(),
-            events: Vec::new(),
-            run_started_ms: started,
-            run_started_instant: Instant::now(),
-            step_started_ms: HashMap::new(),
-            delegation_ids: HashMap::new(),
-            next_delegation_counter: 0,
-        }
-    }
-
-    fn now_ms() -> u128 {
-        SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .map(|d| d.as_millis())
-            .unwrap_or(0)
-    }
-
-    /// Record a phase transition in the runtime lifecycle.
-    pub fn lifecycle_phase_entered(&mut self, phase: RuntimeLifecyclePhase) {
-        let elapsed_ms = self.run_started_instant.elapsed().as_millis();
-        let ts_ms = self.run_started_ms.saturating_add(elapsed_ms);
-        self.events.push(TraceEvent::LifecyclePhaseEntered {
-            ts_ms,
-            elapsed_ms,
-            phase,
-        });
-    }
-
-    /// Record a boundary transition such as runtime init, workflow call, or pause.
-    pub fn execution_boundary_crossed(&mut self, boundary: ExecutionBoundary, state: &str) {
-        let elapsed_ms = self.run_started_instant.elapsed().as_millis();
-        let ts_ms = self.run_started_ms.saturating_add(elapsed_ms);
-        self.events.push(TraceEvent::ExecutionBoundaryCrossed {
-            ts_ms,
-            elapsed_ms,
-            boundary,
-            state: state.to_string(),
-        });
-    }
-
-    pub fn governed_proposal_observed(
-        &mut self,
-        proposal_id: &str,
-        tool_name: &str,
-        redacted_arguments_ref: &str,
-    ) {
-        let elapsed_ms = self.run_started_instant.elapsed().as_millis();
-        let ts_ms = self.run_started_ms.saturating_add(elapsed_ms);
-        self.events.push(TraceEvent::GovernedProposalObserved {
-            ts_ms,
-            elapsed_ms,
-            proposal_id: proposal_id.to_string(),
-            tool_name: tool_name.to_string(),
-            redacted_arguments_ref: redacted_arguments_ref.to_string(),
-        });
-    }
-
-    pub fn governed_proposal_normalized(
-        &mut self,
-        proposal_id: &str,
-        normalized_proposal_ref: &str,
-        redacted_arguments_ref: &str,
-    ) {
-        let elapsed_ms = self.run_started_instant.elapsed().as_millis();
-        let ts_ms = self.run_started_ms.saturating_add(elapsed_ms);
-        self.events.push(TraceEvent::GovernedProposalNormalized {
-            ts_ms,
-            elapsed_ms,
-            proposal_id: proposal_id.to_string(),
-            normalized_proposal_ref: normalized_proposal_ref.to_string(),
-            redacted_arguments_ref: redacted_arguments_ref.to_string(),
-        });
-    }
-
-    pub fn governed_acc_constructed(
-        &mut self,
-        proposal_id: &str,
-        acc_contract_id: &str,
-        replay_posture: &str,
-    ) {
-        let elapsed_ms = self.run_started_instant.elapsed().as_millis();
-        let ts_ms = self.run_started_ms.saturating_add(elapsed_ms);
-        self.events.push(TraceEvent::GovernedAccConstructed {
-            ts_ms,
-            elapsed_ms,
-            proposal_id: proposal_id.to_string(),
-            acc_contract_id: acc_contract_id.to_string(),
-            replay_posture: replay_posture.to_string(),
-        });
-    }
-
-    pub fn governed_policy_injected(
-        &mut self,
-        proposal_id: &str,
-        policy_evidence_ref: &str,
-        outcome: &str,
-    ) {
-        let elapsed_ms = self.run_started_instant.elapsed().as_millis();
-        let ts_ms = self.run_started_ms.saturating_add(elapsed_ms);
-        self.events.push(TraceEvent::GovernedPolicyInjected {
-            ts_ms,
-            elapsed_ms,
-            proposal_id: proposal_id.to_string(),
-            policy_evidence_ref: policy_evidence_ref.to_string(),
-            outcome: outcome.to_string(),
-        });
-    }
-
-    pub fn governed_visibility_resolved(
-        &mut self,
-        proposal_id: &str,
-        actor_view: &str,
-        operator_view: &str,
-        reviewer_view: &str,
-        public_report_view: &str,
-        observatory_projection: &str,
-    ) {
-        let elapsed_ms = self.run_started_instant.elapsed().as_millis();
-        let ts_ms = self.run_started_ms.saturating_add(elapsed_ms);
-        self.events.push(TraceEvent::GovernedVisibilityResolved {
-            ts_ms,
-            elapsed_ms,
-            proposal_id: proposal_id.to_string(),
-            actor_view: sanitize_governed_text(actor_view),
-            operator_view: sanitize_governed_text(operator_view),
-            reviewer_view: sanitize_governed_text(reviewer_view),
-            public_report_view: sanitize_governed_text(public_report_view),
-            observatory_projection: sanitize_governed_text(observatory_projection),
-        });
-    }
-
-    pub fn governed_freedom_gate_decided(
-        &mut self,
-        proposal_id: &str,
-        candidate_id: &str,
-        decision: &str,
-        reason_code: &str,
-        boundary: &str,
-        redaction_summary: &str,
-    ) {
-        let elapsed_ms = self.run_started_instant.elapsed().as_millis();
-        let ts_ms = self.run_started_ms.saturating_add(elapsed_ms);
-        self.events.push(TraceEvent::GovernedFreedomGateDecided {
-            ts_ms,
-            elapsed_ms,
-            proposal_id: proposal_id.to_string(),
-            candidate_id: candidate_id.to_string(),
-            decision: decision.to_string(),
-            reason_code: reason_code.to_string(),
-            boundary: boundary.to_string(),
-            redaction_summary: sanitize_governed_text(redaction_summary),
-        });
-    }
-
-    pub fn governed_action_selected(
-        &mut self,
-        proposal_id: &str,
-        action_id: &str,
-        tool_name: &str,
-        adapter_id: &str,
-        evidence_refs: Vec<String>,
-    ) {
-        let elapsed_ms = self.run_started_instant.elapsed().as_millis();
-        let ts_ms = self.run_started_ms.saturating_add(elapsed_ms);
-        self.events.push(TraceEvent::GovernedActionSelected {
-            ts_ms,
-            elapsed_ms,
-            proposal_id: proposal_id.to_string(),
-            action_id: action_id.to_string(),
-            tool_name: tool_name.to_string(),
-            adapter_id: adapter_id.to_string(),
-            evidence_refs,
-        });
-    }
-
-    pub fn governed_action_rejected(
-        &mut self,
-        proposal_id: &str,
-        action_id: &str,
-        tool_name: &str,
-        adapter_id: &str,
-        reason_code: &str,
-        evidence_refs: Vec<String>,
-    ) {
-        let elapsed_ms = self.run_started_instant.elapsed().as_millis();
-        let ts_ms = self.run_started_ms.saturating_add(elapsed_ms);
-        self.events.push(TraceEvent::GovernedActionRejected {
-            ts_ms,
-            elapsed_ms,
-            proposal_id: proposal_id.to_string(),
-            action_id: action_id.to_string(),
-            tool_name: tool_name.to_string(),
-            adapter_id: adapter_id.to_string(),
-            reason_code: reason_code.to_string(),
-            evidence_refs,
-        });
-    }
-
-    pub fn governed_execution_result(
-        &mut self,
-        proposal_id: &str,
-        action_id: &str,
-        adapter_id: &str,
-        result_ref: &str,
-        evidence_refs: Vec<String>,
-    ) {
-        let elapsed_ms = self.run_started_instant.elapsed().as_millis();
-        let ts_ms = self.run_started_ms.saturating_add(elapsed_ms);
-        self.events
-            .push(TraceEvent::GovernedExecutionResultRecorded {
-                ts_ms,
-                elapsed_ms,
-                proposal_id: proposal_id.to_string(),
-                action_id: action_id.to_string(),
-                adapter_id: adapter_id.to_string(),
-                result_ref: result_ref.to_string(),
-                evidence_refs,
-            });
-    }
-
-    pub fn governed_refusal(
-        &mut self,
-        proposal_id: &str,
-        action_id: &str,
-        reason_code: &str,
-        evidence_refs: Vec<String>,
-    ) {
-        let elapsed_ms = self.run_started_instant.elapsed().as_millis();
-        let ts_ms = self.run_started_ms.saturating_add(elapsed_ms);
-        self.events.push(TraceEvent::GovernedRefusalRecorded {
-            ts_ms,
-            elapsed_ms,
-            proposal_id: proposal_id.to_string(),
-            action_id: action_id.to_string(),
-            reason_code: reason_code.to_string(),
-            evidence_refs,
-        });
-    }
-
-    pub fn governed_redaction_decision(
-        &mut self,
-        proposal_id: &str,
-        audience: &str,
-        surfaces: Vec<String>,
-        outcome: &str,
-        detail: Option<&str>,
-    ) {
-        let elapsed_ms = self.run_started_instant.elapsed().as_millis();
-        let ts_ms = self.run_started_ms.saturating_add(elapsed_ms);
-        self.events
-            .push(TraceEvent::GovernedRedactionDecisionRecorded {
-                ts_ms,
-                elapsed_ms,
-                proposal_id: proposal_id.to_string(),
-                audience: audience.to_string(),
-                surfaces,
-                outcome: outcome.to_string(),
-                detail: detail.map(sanitize_governed_text),
-            });
-    }
-
-    /// Mark a step start with normalized actor/provider/task metadata.
-    pub fn step_started(
-        &mut self,
-        step_id: &str,
-        agent_id: &str,
-        provider_id: &str,
-        task_id: &str,
-        delegation: Option<&DelegationSpec>,
-    ) {
-        let elapsed_ms = self.run_started_instant.elapsed().as_millis();
-        let ts_ms = self.run_started_ms.saturating_add(elapsed_ms);
-        self.events.push(TraceEvent::StepStarted {
-            ts_ms,
-            elapsed_ms,
-            step_id: step_id.to_string(),
-            agent_id: agent_id.to_string(),
-            provider_id: provider_id.to_string(),
-            task_id: task_id.to_string(),
-            delegation: delegation.cloned(),
-        });
-        self.step_started_ms.insert(step_id.to_string(), elapsed_ms);
-    }
-
-    pub fn run_failed(&mut self, message: &str) {
-        let elapsed_ms = self.run_started_instant.elapsed().as_millis();
-        let ts_ms = self.run_started_ms.saturating_add(elapsed_ms);
-        self.events.push(TraceEvent::RunFailed {
-            ts_ms,
-            elapsed_ms,
-            message: message.to_string(),
-        });
-    }
-
-    pub fn scheduler_policy(&mut self, max_concurrency: usize, source: &str) {
-        let elapsed_ms = self.run_started_instant.elapsed().as_millis();
-        let ts_ms = self.run_started_ms.saturating_add(elapsed_ms);
-        self.events.push(TraceEvent::SchedulerPolicy {
-            ts_ms,
-            elapsed_ms,
-            max_concurrency,
-            source: source.to_string(),
-        });
-    }
-
-    pub fn prompt_assembled(&mut self, step_id: &str, prompt_hash: &str) {
-        let elapsed_ms = self.run_started_instant.elapsed().as_millis();
-        let ts_ms = self.run_started_ms.saturating_add(elapsed_ms);
-        self.events.push(TraceEvent::PromptAssembled {
-            ts_ms,
-            elapsed_ms,
-            step_id: step_id.to_string(),
-            prompt_hash: prompt_hash.to_string(),
-        });
-    }
-
-    pub fn step_output_chunk(&mut self, step_id: &str, chunk_bytes: usize) {
-        let elapsed_ms = self.run_started_instant.elapsed().as_millis();
-        let ts_ms = self.run_started_ms.saturating_add(elapsed_ms);
-        self.events.push(TraceEvent::StepOutputChunk {
-            ts_ms,
-            elapsed_ms,
-            step_id: step_id.to_string(),
-            chunk_bytes,
-        });
-    }
-
-    fn delegation_id_for_step(&mut self, step_id: &str) -> String {
-        if let Some(existing) = self.delegation_ids.get(step_id) {
-            return existing.clone();
-        }
-        self.next_delegation_counter = self.next_delegation_counter.saturating_add(1);
-        let delegation_id = format!("del-{}", self.next_delegation_counter);
-        self.delegation_ids
-            .insert(step_id.to_string(), delegation_id.clone());
-        delegation_id
-    }
-
-    pub fn delegation_requested(&mut self, step_id: &str, action_kind: &str, target_id: &str) {
-        let elapsed_ms = self.run_started_instant.elapsed().as_millis();
-        let ts_ms = self.run_started_ms.saturating_add(elapsed_ms);
-        let delegation_id = self.delegation_id_for_step(step_id);
-        self.events.push(TraceEvent::DelegationRequested {
-            ts_ms,
-            elapsed_ms,
-            delegation_id,
-            step_id: step_id.to_string(),
-            action_kind: action_kind.to_string(),
-            target_id: target_id.to_string(),
-        });
-    }
-
-    pub fn delegation_policy_evaluated(
-        &mut self,
-        step_id: &str,
-        action_kind: &str,
-        target_id: &str,
-        decision: &str,
-        rule_id: Option<&str>,
-    ) {
-        let elapsed_ms = self.run_started_instant.elapsed().as_millis();
-        let ts_ms = self.run_started_ms.saturating_add(elapsed_ms);
-        let delegation_id = self.delegation_id_for_step(step_id);
-        self.events.push(TraceEvent::DelegationPolicyEvaluated {
-            ts_ms,
-            elapsed_ms,
-            delegation_id,
-            step_id: step_id.to_string(),
-            action_kind: action_kind.to_string(),
-            target_id: target_id.to_string(),
-            decision: decision.to_string(),
-            rule_id: rule_id.map(|v| v.to_string()),
-        });
-    }
-
-    pub fn delegation_approved(&mut self, step_id: &str) {
-        let elapsed_ms = self.run_started_instant.elapsed().as_millis();
-        let ts_ms = self.run_started_ms.saturating_add(elapsed_ms);
-        let delegation_id = self.delegation_id_for_step(step_id);
-        self.events.push(TraceEvent::DelegationApproved {
-            ts_ms,
-            elapsed_ms,
-            delegation_id,
-            step_id: step_id.to_string(),
-        });
-    }
-
-    pub fn delegation_denied(
-        &mut self,
-        step_id: &str,
-        action_kind: &str,
-        target_id: &str,
-        rule_id: Option<&str>,
-    ) {
-        let elapsed_ms = self.run_started_instant.elapsed().as_millis();
-        let ts_ms = self.run_started_ms.saturating_add(elapsed_ms);
-        let delegation_id = self.delegation_id_for_step(step_id);
-        self.events.push(TraceEvent::DelegationDenied {
-            ts_ms,
-            elapsed_ms,
-            delegation_id,
-            step_id: step_id.to_string(),
-            action_kind: action_kind.to_string(),
-            target_id: target_id.to_string(),
-            rule_id: rule_id.map(|v| v.to_string()),
-        });
-    }
-
-    pub fn delegation_dispatched(&mut self, step_id: &str, action_kind: &str, target_id: &str) {
-        let elapsed_ms = self.run_started_instant.elapsed().as_millis();
-        let ts_ms = self.run_started_ms.saturating_add(elapsed_ms);
-        let delegation_id = self.delegation_id_for_step(step_id);
-        self.events.push(TraceEvent::DelegationDispatched {
-            ts_ms,
-            elapsed_ms,
-            delegation_id,
-            step_id: step_id.to_string(),
-            action_kind: action_kind.to_string(),
-            target_id: target_id.to_string(),
-        });
-    }
-
-    pub fn delegation_result_received(
-        &mut self,
-        step_id: &str,
-        success: bool,
-        output_bytes: usize,
-    ) {
-        let elapsed_ms = self.run_started_instant.elapsed().as_millis();
-        let ts_ms = self.run_started_ms.saturating_add(elapsed_ms);
-        let delegation_id = self.delegation_id_for_step(step_id);
-        self.events.push(TraceEvent::DelegationResultReceived {
-            ts_ms,
-            elapsed_ms,
-            delegation_id,
-            step_id: step_id.to_string(),
-            success,
-            output_bytes,
-        });
-    }
-
-    pub fn delegation_completed(&mut self, step_id: &str, outcome: &str) {
-        let elapsed_ms = self.run_started_instant.elapsed().as_millis();
-        let ts_ms = self.run_started_ms.saturating_add(elapsed_ms);
-        let delegation_id = self.delegation_id_for_step(step_id);
-        self.events.push(TraceEvent::DelegationCompleted {
-            ts_ms,
-            elapsed_ms,
-            delegation_id,
-            step_id: step_id.to_string(),
-            outcome: outcome.to_string(),
-        });
-    }
-
-    /// Record a terminal step result including success and duration.
-    pub fn step_finished(&mut self, step_id: &str, success: bool) {
-        let elapsed_ms = self.run_started_instant.elapsed().as_millis();
-        let ts_ms = self.run_started_ms.saturating_add(elapsed_ms);
-        let duration_ms = self
-            .step_started_ms
-            .remove(step_id)
-            .map(|started| elapsed_ms.saturating_sub(started))
-            .unwrap_or(0);
-        self.events.push(TraceEvent::StepFinished {
-            ts_ms,
-            elapsed_ms,
-            step_id: step_id.to_string(),
-            success,
-            duration_ms,
-        });
-    }
-
-    pub fn call_entered(
-        &mut self,
-        caller_step_id: &str,
-        callee_workflow_id: &str,
-        namespace: &str,
-    ) {
-        self.execution_boundary_crossed(ExecutionBoundary::WorkflowCall, "entered");
-        let elapsed_ms = self.run_started_instant.elapsed().as_millis();
-        let ts_ms = self.run_started_ms.saturating_add(elapsed_ms);
-        self.events.push(TraceEvent::CallEntered {
-            ts_ms,
-            elapsed_ms,
-            caller_step_id: caller_step_id.to_string(),
-            callee_workflow_id: callee_workflow_id.to_string(),
-            namespace: namespace.to_string(),
-        });
-    }
-
-    pub fn call_exited(&mut self, caller_step_id: &str, status: &str, namespace: &str) {
-        self.execution_boundary_crossed(ExecutionBoundary::WorkflowCall, status);
-        let elapsed_ms = self.run_started_instant.elapsed().as_millis();
-        let ts_ms = self.run_started_ms.saturating_add(elapsed_ms);
-        self.events.push(TraceEvent::CallExited {
-            ts_ms,
-            elapsed_ms,
-            caller_step_id: caller_step_id.to_string(),
-            status: status.to_string(),
-            namespace: namespace.to_string(),
-        });
-    }
-
-    /// Record final run completion.
-    pub fn run_finished(&mut self, success: bool) {
-        let elapsed_ms = self.run_started_instant.elapsed().as_millis();
-        let ts_ms = self.run_started_ms.saturating_add(elapsed_ms);
-        self.events.push(TraceEvent::RunFinished {
-            ts_ms,
-            elapsed_ms,
-            success,
-        });
-    }
-
-    /// Return elapsed milliseconds from trace start to now.
-    pub fn current_elapsed_ms(&self) -> u128 {
-        self.run_started_instant.elapsed().as_millis()
-    }
-
-    /// Return the current trace timestamp in wall-clock milliseconds.
-    pub fn current_ts_ms(&self) -> u128 {
-        self.run_started_ms
-            .saturating_add(self.run_started_instant.elapsed().as_millis())
-    }
-}
-
-/// Print a human-readable trace to stdout (stable + diff-friendly).
-pub fn print_trace(tr: &Trace) {
-    println!(
-        "TRACE run_id={} workflow_id={} version={}",
-        tr.run_id, tr.workflow_id, tr.version
-    );
-    for ev in &tr.events {
-        println!("{}", ev.summarize());
-    }
-}
-
-fn delegation_json(delegation: Option<&DelegationSpec>) -> Option<String> {
-    let d = delegation?;
-    if d.is_effectively_empty() {
-        return None;
-    }
-    serde_json::to_string(&d.canonicalized()).ok()
-}
-
-fn sanitize_governed_text(value: &str) -> String {
-    if value.contains("/Users/")
-        || value.contains("/home/")
-        || value.contains("sk-")
-        || value.contains("gho_")
-        || value.contains("BEGIN PRIVATE KEY")
-    {
-        return "[redacted-sensitive-text]".to_string();
-    }
-    if value.contains('{') || value.contains('}') {
-        return "[redacted-structured-text]".to_string();
-    }
-    value.to_string()
-}
-
-pub fn format_iso_utc_ms(ts_ms: u128) -> String {
-    format_ts_ms(ts_ms)
-}
-
-fn format_ts_ms(ts_ms: u128) -> String {
-    // Convert unix epoch millis to UTC without external dependencies.
-    // Algorithm adapted from civil calendar conversion by Howard Hinnant.
-    let total_secs = i128::try_from(ts_ms / 1000).unwrap_or(i128::MAX);
-    let millis = ts_ms % 1000;
-
-    let days = total_secs.div_euclid(86_400);
-    let secs_of_day = total_secs.rem_euclid(86_400);
-
-    let z = days + 719_468;
-    let era = if z >= 0 { z } else { z - 146_096 }.div_euclid(146_097);
-    let doe = z - era * 146_097;
-    let yoe = (doe - doe / 1_460 + doe / 36_524 - doe / 146_096).div_euclid(365);
-    let y = yoe + era * 400;
-    let doy = doe - (365 * yoe + yoe / 4 - yoe / 100);
-    let mp = (5 * doy + 2).div_euclid(153);
-    let d = doy - (153 * mp + 2).div_euclid(5) + 1;
-    let m = mp + if mp < 10 { 3 } else { -9 };
-    let year = y + if m <= 2 { 1 } else { 0 };
-
-    let hour = secs_of_day.div_euclid(3_600);
-    let minute = secs_of_day.rem_euclid(3_600).div_euclid(60);
-    let second = secs_of_day.rem_euclid(60);
-
-    format!(
-        "{:04}-{:02}-{:02}T{:02}:{:02}:{:02}.{:03}Z",
-        year, m, d, hour, minute, second, millis
-    )
-}
-
-fn format_duration_secs(duration_ms: u128) -> String {
-    let secs = duration_ms as f64 / 1000.0;
-    format!("{secs:.3}s")
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn trace_records_step_lifecycle_events_in_order() {
-        let mut tr = Trace::new("run-1", "workflow-1", "0.1");
-
-        tr.step_started("step-1", "agent-1", "provider-1", "task-1", None);
-        tr.prompt_assembled("step-1", "hash-123");
-        tr.step_output_chunk("step-1", 5);
-        tr.step_finished("step-1", true);
-
-        assert_eq!(tr.events.len(), 4);
-
-        match &tr.events[0] {
-            TraceEvent::StepStarted { step_id, .. } => {
-                assert_eq!(step_id, "step-1");
-            }
-            _ => panic!("expected StepStarted event"),
-        }
-
-        match &tr.events[1] {
-            TraceEvent::PromptAssembled { prompt_hash, .. } => {
-                assert_eq!(prompt_hash, "hash-123");
-            }
-            _ => panic!("expected PromptAssembled event"),
-        }
-
-        match &tr.events[2] {
-            TraceEvent::StepOutputChunk { chunk_bytes, .. } => {
-                assert_eq!(*chunk_bytes, 5);
-            }
-            _ => panic!("expected StepOutputChunk event"),
-        }
-
-        match &tr.events[3] {
-            TraceEvent::StepFinished {
-                success,
-                duration_ms,
-                ..
-            } => {
-                assert!(*success);
-                assert!(*duration_ms <= 1_000);
-            }
-            _ => panic!("expected StepFinished event"),
-        }
-    }
-
-    #[test]
-    fn trace_records_delegation_denied_event_with_optional_rule_id() {
-        let mut tr = Trace::new("run-1", "wf-1", "0.7");
-        tr.delegation_denied("s1", "provider_call", "local", Some("deny-local"));
-        tr.delegation_denied("s2", "provider_call", "backup", None);
-
-        let line0 = tr.events[0].summarize();
-        let line1 = tr.events[1].summarize();
-        assert!(line0.contains("DelegationDenied"));
-        assert!(line0.contains("action=provider_call"));
-        assert!(line0.contains("target=local"));
-        assert!(line0.contains("rule_id=deny-local"));
-        assert!(line1.contains("DelegationDenied"));
-        assert!(line1.contains("action=provider_call"));
-        assert!(line1.contains("target=backup"));
-        assert!(!line1.contains("rule_id="));
-    }
-
-    #[test]
-    fn trace_records_scheduler_policy_event() {
-        let mut tr = Trace::new("run-1", "wf-1", "0.7");
-        tr.scheduler_policy(4, "engine_default");
-        assert_eq!(tr.events.len(), 1);
-        match &tr.events[0] {
-            TraceEvent::SchedulerPolicy {
-                max_concurrency,
-                source,
-                ..
-            } => {
-                assert_eq!(*max_concurrency, 4);
-                assert_eq!(source, "engine_default");
-                let summary = tr.events[0].summarize();
-                assert!(summary.contains("SchedulerPolicy"));
-                assert!(summary.contains("max_concurrency=4"));
-                assert!(summary.contains("source=engine_default"));
-            }
-            _ => panic!("expected SchedulerPolicy event"),
-        }
-    }
-
-    #[test]
-    fn trace_records_delegation_policy_event_with_optional_rule_id() {
-        let mut tr = Trace::new("run-del-policy", "wf-del-policy", "0.7");
-        tr.delegation_policy_evaluated(
-            "s1",
-            "provider_call",
-            "local",
-            "denied",
-            Some("deny-local"),
-        );
-        tr.delegation_policy_evaluated("s2", "provider_call", "backup", "allowed", None);
-        assert_eq!(tr.events.len(), 2);
-
-        let line0 = tr.events[0].summarize();
-        assert!(line0.contains("DelegationPolicyEvaluated"));
-        assert!(line0.contains("action=provider_call"));
-        assert!(line0.contains("target=local"));
-        assert!(line0.contains("decision=denied"));
-        assert!(line0.contains("rule_id=deny-local"));
-
-        let line1 = tr.events[1].summarize();
-        assert!(line1.contains("decision=allowed"));
-        assert!(!line1.contains("rule_id="));
-    }
-
-    #[test]
-    fn trace_preserves_run_and_workflow_ids() {
-        let tr = Trace::new("run-x", "workflow-y", "0.1");
-
-        assert_eq!(tr.run_id, "run-x");
-        assert_eq!(tr.workflow_id, "workflow-y");
-    }
-
-    #[test]
-    fn trace_allows_multiple_steps() {
-        let mut tr = Trace::new("run-2", "workflow-2", "0.1");
-
-        tr.step_started("step-a", "agent-a", "provider-a", "task-a", None);
-        tr.step_finished("step-a", true);
-
-        tr.step_started("step-b", "agent-b", "provider-b", "task-b", None);
-        tr.step_finished("step-b", false);
-
-        assert_eq!(tr.events.len(), 4);
-    }
-
-    #[test]
-    fn trace_records_call_events() {
-        let mut tr = Trace::new("run-3", "workflow-3", "0.5");
-        tr.call_entered("parent", "child", "ns");
-        tr.call_exited("parent", "success", "ns");
-        assert_eq!(tr.events.len(), 4);
-    }
-
-    #[test]
-    fn trace_records_runtime_lifecycle_and_boundary_events() {
-        let mut tr = Trace::new("run-life", "workflow-life", "0.87.1");
-        tr.lifecycle_phase_entered(RuntimeLifecyclePhase::Init);
-        tr.lifecycle_phase_entered(RuntimeLifecyclePhase::Execute);
-        tr.execution_boundary_crossed(ExecutionBoundary::RuntimeInit, "fresh_start");
-        tr.execution_boundary_crossed(ExecutionBoundary::RunCompletion, "success");
-        tr.lifecycle_phase_entered(RuntimeLifecyclePhase::Complete);
-        tr.lifecycle_phase_entered(RuntimeLifecyclePhase::Teardown);
-
-        let lines: Vec<String> = tr.events.iter().map(TraceEvent::summarize).collect();
-        assert!(lines
-            .iter()
-            .any(|line| line.contains("LifecyclePhaseEntered phase=init")));
-        assert!(lines
-            .iter()
-            .any(|line| line.contains("LifecyclePhaseEntered phase=execute")));
-        assert!(lines.iter().any(|line| line
-            .contains("ExecutionBoundaryCrossed boundary=runtime_init state=fresh_start")));
-        assert!(lines
-            .iter()
-            .any(|line| line
-                .contains("ExecutionBoundaryCrossed boundary=run_completion state=success")));
-        assert!(lines
-            .iter()
-            .any(|line| line.contains("LifecyclePhaseEntered phase=complete")));
-        assert!(lines
-            .iter()
-            .any(|line| line.contains("LifecyclePhaseEntered phase=teardown")));
-    }
-
-    #[test]
-    fn delegation_ids_are_stable_per_step_and_increment_in_order() {
-        let mut tr = Trace::new("run-del", "workflow-del", "0.7");
-        tr.delegation_requested("s1", "provider_call", "local");
-        tr.delegation_policy_evaluated("s1", "provider_call", "local", "allowed", None);
-        tr.delegation_requested("s2", "remote_exec", "remote");
-
-        let ids: Vec<String> = tr
-            .events
-            .iter()
-            .filter_map(|event| match event {
-                TraceEvent::DelegationRequested { delegation_id, .. } => {
-                    Some(delegation_id.clone())
-                }
-                TraceEvent::DelegationPolicyEvaluated { delegation_id, .. } => {
-                    Some(delegation_id.clone())
-                }
-                _ => None,
-            })
-            .collect();
-        assert_eq!(ids, vec!["del-1", "del-1", "del-2"]);
-    }
-
-    #[test]
-    fn delegation_event_summaries_are_stable_and_safe() {
-        let mut tr = Trace::new("run-del", "workflow-del", "0.7");
-        tr.delegation_requested("s1", "provider_call", "local");
-        tr.delegation_policy_evaluated("s1", "provider_call", "local", "allowed", None);
-        tr.delegation_dispatched("s1", "provider_call", "local");
-        tr.delegation_result_received("s1", true, 12);
-        tr.delegation_completed("s1", "success");
-
-        let lines: Vec<String> = tr.events.iter().map(TraceEvent::summarize).collect();
-        assert!(lines.iter().any(|line| line.contains(
-            "DelegationRequested delegation_id=del-1 step=s1 action=provider_call target=local"
-        )));
-        assert!(lines.iter().any(|line| line.contains(
-            "DelegationPolicyEvaluated delegation_id=del-1 step=s1 action=provider_call target=local decision=allowed"
-        )));
-        assert!(lines.iter().any(|line| line.contains(
-            "DelegationDispatched delegation_id=del-1 step=s1 action=provider_call target=local"
-        )));
-        assert!(lines.iter().any(|line| line.contains(
-            "DelegationResultReceived delegation_id=del-1 step=s1 success=true bytes=12"
-        )));
-        assert!(lines
-            .iter()
-            .any(|line| line
-                .contains("DelegationCompleted delegation_id=del-1 step=s1 outcome=success")));
-    }
-
-    #[test]
-    fn trace_step_started_includes_delegation_when_present() {
-        let mut tr = Trace::new("run-del", "workflow-del", "0.6");
-        tr.step_started(
-            "step-del",
-            "agent-del",
-            "provider-del",
-            "task-del",
-            Some(&DelegationSpec {
-                role: Some("reviewer".to_string()),
-                requires_verification: Some(true),
-                escalation_target: Some("human".to_string()),
-                tags: vec!["safety".to_string(), "compliance".to_string()],
-            }),
-        );
-        let line = tr.events[0].summarize();
-        assert!(
-            line.contains("delegation={\"role\":\"reviewer\",\"requires_verification\":true,\"escalation_target\":\"human\",\"tags\":[\"compliance\",\"safety\"]}"),
-            "line was:\n{line}"
-        );
-    }
-
-    #[test]
-    fn trace_records_governed_execution_events_without_payload_leakage() {
-        let mut tr = Trace::new("run-governed", "workflow-governed", "0.90.5");
-        tr.governed_proposal_observed(
-            "proposal.fixture.safe-read",
-            "fixture.safe_read",
-            "artifacts/run-governed/governed/proposal_arguments.redacted.json",
-        );
-        tr.governed_proposal_normalized(
-            "proposal.fixture.safe-read",
-            "normalized.proposal.fixture.safe-read",
-            "artifacts/run-governed/governed/proposal_arguments.redacted.json",
-        );
-        tr.governed_acc_constructed(
-            "proposal.fixture.safe-read",
-            "acc.compiler.proposal.fixture.safe-read",
-            "deterministic_fixture_compiler",
-        );
-        tr.governed_policy_injected(
-            "proposal.fixture.safe-read",
-            "policy.fixture.safe-read",
-            "allowed",
-        );
-        tr.governed_visibility_resolved(
-            "proposal.fixture.safe-read",
-            "compiled ACC request status",
-            "full compiler fixture evidence",
-            "redacted compiler evidence and policy result",
-            "aggregate compiler pass/fail only",
-            "redacted compiler governance event",
-        );
-        tr.governed_freedom_gate_decided(
-            "proposal.fixture.safe-read",
-            "candidate.safe-read",
-            "allowed",
-            "gate_allowed",
-            "execution",
-            "private_arguments_redacted digest=sha256:bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb",
-        );
-        tr.governed_action_selected(
-            "proposal.fixture.safe-read",
-            "action.safe_read",
-            "fixture.safe_read",
-            "adapter.fixture.safe_read.dry_run",
-            vec![
-                "proposal:proposal.fixture.safe-read".to_string(),
-                "acc:acc.compiler.proposal.fixture.safe-read".to_string(),
-                "action:fixture_read".to_string(),
-                "normalized_proposal:normalized.proposal.fixture.safe-read".to_string(),
-                "policy:policy.fixture.safe-read".to_string(),
-                "gate:candidate.safe-read".to_string(),
-                "execution:action.safe_read".to_string(),
-            ],
-        );
-        tr.governed_execution_result(
-            "proposal.fixture.safe-read",
-            "action.safe_read",
-            "adapter.fixture.safe_read.dry_run",
-            "artifacts/run-governed/governed/result.redacted.json",
-            vec![
-                "gate:candidate.safe-read".to_string(),
-                "execution:action.safe_read".to_string(),
-            ],
-        );
-        tr.governed_redaction_decision(
-            "proposal.fixture.safe-read",
-            "reviewer",
-            vec![
-                "arguments".to_string(),
-                "results".to_string(),
-                "errors".to_string(),
-                "rejected_alternatives".to_string(),
-            ],
-            "redacted",
-            Some("digest_only"),
-        );
-
-        let lines: Vec<String> = tr.events.iter().map(TraceEvent::summarize).collect();
-        assert!(lines
-            .iter()
-            .any(|line| line.contains("GovernedProposalObserved")));
-        assert!(lines
-            .iter()
-            .any(|line| line.contains("GovernedFreedomGateDecided")));
-        assert!(lines
-            .iter()
-            .any(|line| line.contains("GovernedExecutionResultRecorded")));
-        assert!(!lines
-            .iter()
-            .any(|line| line.contains("fixture-secret-token")));
-    }
-
-    #[test]
-    fn trace_sanitizes_sensitive_governed_text_fields() {
-        let mut tr = Trace::new("run-governed", "workflow-governed", "0.90.5");
-        tr.governed_visibility_resolved(
-            "proposal.fixture.safe-read",
-            "compiled ACC request status",
-            "{\"secret\":\"leak\"}",
-            "/Users/test/private.txt",
-            "aggregate compiler pass/fail only",
-            "sk-secret-token",
-        );
-        tr.governed_redaction_decision(
-            "proposal.fixture.safe-read",
-            "reviewer",
-            vec!["arguments".to_string()],
-            "redacted",
-            Some("gho_sensitive"),
-        );
-
-        let lines: Vec<String> = tr.events.iter().map(TraceEvent::summarize).collect();
-        assert!(lines
-            .iter()
-            .any(|line| line.contains("[redacted-sensitive-text]")));
-        assert!(lines
-            .iter()
-            .any(|line| line.contains("[redacted-structured-text]")));
-        assert!(!lines
-            .iter()
-            .any(|line| line.contains("/Users/test/private.txt")));
-        assert!(!lines.iter().any(|line| line.contains("gho_sensitive")));
-    }
-}
